Revert "[Collocation] Add Monitor Table Functionality"
