+++ platform/package.json
@@ -17,8 +17,6 @@
     "@tailwindcss/typography": "^0.5.9",
     "autoprefixer": "^10.4.12",
     "daisyui": "^2.47.0",
-    "date-fns": "^2.29.3",
-    "dayjs": "^1.11.7",
     "flowbite": "^1.6.3",
     "flowbite-react": "^0.3.8",
     "moment": "^2.29.4",
@@ -27,10 +25,8 @@
     "next-redux-wrapper": "^8.1.0",
     "postcss": "^8.4.16",
     "react": "^18.2.0",
-    "react-calendar": "^4.0.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
-    "react-tailwindcss-datepicker": "^1.4.2",
     "sass": "^1.54.3",
     "tailwindcss": "^3.1.8"
   },
