+++ platform/src/common/components/Button/index.jsx
@@ -1,25 +1,13 @@
 const Button = ({ onClick, className, path, children }) => {
-  if (path) {
-    // if the button has an href property, it should be a link button
-    return (
-      <a
-        href={path}
-        className={`flex justify-center items-center normal-case px-4 py-2 ${className}`}
-      >
-        {children}
-      </a>
-    );
-  } else {
-    // otherwise, it's a regular button with an onClick handler
-    return (
-      <button
-        onClick={onClick}
-        className={`flex justify-center items-center normal-case px-4 py-2 ${className}`}
-      >
-        {children}
-      </button>
-    );
-  }
+  return (
+    <a
+      href={path || '#'}
+      onClick={onClick}
+      className={`flex justify-center items-center normal-case px-4 py-2 ${className}`}
+    >
+      {children}
+    </a>
+  );
 };
 
 export default Button;
