--- platform/src/common/components/Layout/content_box.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const ContentBox = ({ children }) => {
-  return (
-    <div className='mx-6 mb-6 border-[0.5px] rounded-lg border-grey-150 flex items-center'>
-      {children}
-    </div>
-  );
-};
-
-export default ContentBox;
