+++ platform/src/common/components/Skeleton/index.jsx
@@ -1,0 +1,5 @@
+const Skeleton = ({ width }) => {
+  return <div role='status' class='max-w-sm animate-pulse'></div>;
+};
+
+export default Skeleton;
