--- platform/src/core/urls/deviceMonitoring.js
@@ -1,5 +1,0 @@
-import { stripTrailingSlash } from '../utils/strings';
-
-const BASE_DEVICE_MONITORING_URL = stripTrailingSlash(process.env.NEXT_PUBLIC_API_BASE_URL);
-
-export const COLLOCATION = `${BASE_DEVICE_MONITORING_URL}/monitor/devices/collocation`;
