+++ platform/src/lib/store/index.js
@@ -1,20 +1,14 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { createWrapper } from 'next-redux-wrapper';
 import { deviceRegistryApi } from './services/deviceRegistry';
-import selectedCollocateDevicesReducer from './services/addMonitor/selectedCollocateDevicesSlice';
-import { collocateApi } from './services/collocation';
-import collocationDataReducer from './services/collocation/collocationDataSlice';
 
 const store = () =>
   configureStore({
     reducer: {
       [deviceRegistryApi.reducerPath]: deviceRegistryApi.reducer,
-      [collocateApi.reducerPath]: collocateApi.reducer,
-      selectedCollocateDevices: selectedCollocateDevicesReducer,
-      collocationData: collocationDataReducer,
     },
     middleware: (getDefaultMiddleware) =>
-      getDefaultMiddleware().concat(deviceRegistryApi.middleware, collocateApi.middleware),
+      getDefaultMiddleware().concat(deviceRegistryApi.middleware),
   });
 
 export const wrapper = createWrapper(store);
