+++ platform/tailwind.config.js
@@ -1,11 +1,7 @@
 const defaultTheme = require('tailwindcss/defaultTheme');
 
 module.exports = {
-  content: [
-    './src/pages/**/*.{js,jsx}',
-    './src/common/components/**/*.{js,jsx}',
-    './node_modules/react-tailwindcss-datepicker/dist/index.esm.js',
-  ],
+  content: ['./src/pages/**/*.{js,jsx}', './src/common/components/**/*.{js,jsx}'],
   theme: {
     extend: {
       fontFamily: {
@@ -19,18 +15,8 @@ module.exports = {
         'light-blue': '#135DFF14',
         'sidebar-blue': '#135DFF0A',
         'dark-blue': '#0F4ACC',
-        'light-text': '#6D7175',
-        'baby-blue': '#F5F8FF',
         grey: {
-          100: '#363A4414',
-          150: '#363A4429',
           300: '#6D7175',
-          150: '#363A4429',
-          100: '#363A4414',
-        },
-        skeleton: '#363A4429',
-        black: {
-          600: '#202223',
         },
       },
     },
